Updating NSArray_Extensions for 64-bit compatibility.

+++ NSArray_Extensions.h
@@ -13,6 +13,6 @@
 @end
 
 @interface NSMutableArray (KBExtensions)
-- (void)insertObjectsFromArray:(NSArray *)array atIndex:(int)index;
+- (void)insertObjectsFromArray:(NSArray *)array atIndex:(NSInteger)index;
 @end
